requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
ccxt>=4.0.0  # For market data
backtrader>=1.9.76.123  # Backtesting engine
streamlit>=1.28.0  # For dashboards
plotly>=5.17.0  # Charts
scikit-learn>=1.3.0  # ML models
joblib>=1.3.0  # Model serialization
pytest>=7.4.0  # Testing
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting